Version 1 update
Added options for demultiplexed fastq.gz file.
Added options for explicit quality filter.
Added QoL changes.

+++ CLIPittyClip.sh
@@ -3,6 +3,10 @@
 EXP_ID=""
 # Type of CLIP experiment for naming:
 EXP_TYPE=""
+# Quality filter threshod:
+QUALITY=30
+# Perform or skip demultiplexing:
+DEMUX="no"
 # Allowed mismatch for barcode demultiplexing:
 barcodeMis=0
 # Minimum read length after 5 and 3' trimming/clipping:
@@ -27,17 +31,18 @@ function show_info {
   echo "-----------------------------------------------------------------------------------------------"
   echo "CLIPittyClip: Single-line CLIP data analysis pipeline"
   echo "-----------------------------------------------------------------------------------------------"
-  echo "Version 0.2"
+  echo "Version 1.0.0"
   echo "Author: Soon Yi"
-  echo "Last update: 2023-08-06"
+  echo "Last update: 2023-10-25"
   echo "-----------------------------------------------------------------------------------------------"
   echo "This is a standard CLIP analysis pipeline, from fastq.gz to peaks."
+  echo "If input fastq requires demultiplexing, use -d option and provide proper barcode file."
   echo "The pipeline utilizes following programs: "
   echo " - fastx_toolkit (collapser, barcode_splitter, trimmer, clipper)"
   echo " - ctk (stripBarcode, tag2collapse)"
   echo " - bowtie2"
   echo " - samtools (view, sort, index)"
-  echo " - bedtools (bamtobed, genomecov)"
+  echo " - bedtools (bamtobed, genomecov, coverage)"
   echo " - Homer (makeTagDirectory, findPeaks)"
   echo "Make sure conda environment has all the necessary programs installed."
   echo "-----------------------------------------------------------------------------------------------"
@@ -47,16 +52,29 @@ function show_info {
 function show_usage {
   echo "-----------------------------------------------------------------------------------------------"
   echo "Usage: CLIPittyClip.sh -i ? -y ? ... -x ? ..."
-  echo "- Experiment ID (-i), type (-y) and genome index (-x) are required."
-  echo "- Starting fastq.gz file should have name convention as: id_type.fastq.gz"
+  echo "- Experiment ID (-i), type (-y) and genome index (-x) are required:"
+  echo "  Experiment ID (-i) represents unique identifier for your experiment."
+  echo "  Experiment type (-y) represents short descripter for your experiment."
+  echo "  Starting fastq.gz file should have name convention as: id_type.fastq.gz"
   echo "  Example: JL1000_Input.fastq.gz"
+  echo "-----------------------------------------------------------------------------------------------"
+  echo "- If demultiplexing is required, make sure to set -d option to yes."
+  echo "  Barcode file should have the same name convention as fastq.gz file, followed by '_BC.txt'."
+  echo "  Example: JL1000_Input_BC.txt"
+  echo "- Check our GitHub page for sample barcode file."
+  echo "-----------------------------------------------------------------------------------------------"
   echo "- Path and file name to genome index for bowtie2 mapping (-x) should follow bowtie2 convention."
   echo "  Path/file name will also be used for .fa.fai file for bedtools genomecov."
+  echo "  Makre sure bowtie2 genome index and bedtools genome files exists in your designated path."
+  echo "-----------------------------------------------------------------------------------------------"
   echo "- The other options will use default options if not specified."
   echo "-----------------------------------------------------------------------------------------------"
   echo "Options:"
+  echo "  -h: print usage information"
   echo "  -i: experiment ID"
   echo "  -y: experiment type (e.g., Input, Enrich, or Fraction)"
+  echo "  -q: quality score threshold                             (default: 30)"
+  echo "  -d: barcode file for demultiplexing.                    (default: no)"
   echo "  -b: allowed barcode mismatch for demultiplexing         (default:  0)"
   echo "  -l: minimum read length after 5/3 end trimming/clipping (default: 16)"
   echo "  -x: path to genome index for bowtie2 mapping"
@@ -70,17 +88,31 @@ function show_usage {
 }
 
 function on_exit {
-  read -p "Press Enter to exit."
+  read -p "Press enter to exit."
 }
 
+# Check if arguments are provided:
+if [[ $# -eq 0 ]]; then
+  show_info
+  echo "Error: No option arguments have been passed. use '-h' option for usage information."
+  on_exit
+  exit 1
+fi
+
 # Parse command-line options
-while getopts "i:y:b:l:x:m:s:t:p:z:f" opt; do
+while getopts "i:y:q:d:b:l:x:m:s:t:p:z:f:h" opt; do
   case $opt in
     i)
       EXP_ID="$OPTARG"
       ;;
     y)
       EXP_TYPE="$OPTARG"
+      ;;
+    q)
+      QUALITY="$OPTARG"
+      ;;
+    d)
+      DEMUX="$OPTARG"
       ;;
     b)
       barcodeMis="$OPTARG"
@@ -109,15 +141,19 @@ while getopts "i:y:b:l:x:m:s:t:p:z:f" opt; do
     f)
       FragLength="$OPTARG"
       ;;
-    \?)
+    h)
       show_info
+      show_usage
+      on_exit
+      exit 0
+      ;;
+    \?)
       echo "Error: Invalid option -$OPTARG" >&2
       show_usage
       on_exit
       exit 1
       ;;
     :)
-      show_info
       echo "Error: Option -$OPTARG requires an argument." >&2
       show_usage
       on_exit
@@ -127,40 +163,65 @@ while getopts "i:y:b:l:x:m:s:t:p:z:f" opt; do
 done
 
 # Check if all required options are provided
-if [[ -z "$EXP_ID" || -z "$EXP_TYPE" ]]; then
-  show_info
+if [[ -z "$EXP_ID" || -z "$EXP_TYPE" || -z "$genome_index" ]]; then
   echo "Error: Options -i, -y, and -x are required."
   show_usage
   on_exit
   exit 1
 fi
 
-BASE_NAME="${EXP_ID}_${EXP_TYPE}"
-echo "-----------------------------------------------------------------------------------------------" | tee -a ${BASE_NAME}_logs.txt
-echo "CLIPittyClip: Single-line CLIP data analysis pipeline" | tee -a ${BASE_NAME}_logs.txt
-echo "Version 0.2" | tee -a ${BASE_NAME}_logs.txt
-echo "Author: Soon Yi" | tee -a ${BASE_NAME}_logs.txt
-echo "Last update: 2023-08-06" | tee -a ${BASE_NAME}_logs.txt
-echo "-----------------------------------------------------------------------------------------------" | tee -a ${BASE_NAME}_logs.txt
-echo "This is a standard CLIP analysis pipeline, from fastq.gz to peaks." | tee -a ${BASE_NAME}_logs.txt
-echo "The pipeline utilizes following programs: " | tee -a ${BASE_NAME}_logs.txt
-echo " - fastx_toolkit (collapser, barcode_splitter, trimmer, clipper)" | tee -a ${BASE_NAME}_logs.txt
-echo " - ctk (stripBarcode, tag2collapse)" | tee -a ${BASE_NAME}_logs.txt
-echo " - bowtie2" | tee -a ${BASE_NAME}_logs.txt
-echo " - samtools (view, sort, index)" | tee -a ${BASE_NAME}_logs.txt
-echo " - bedtools (bamtobed, genomecov)" | tee -a ${BASE_NAME}_logs.txt
-echo " - Homer (makeTagDirectory, findPeaks)" | tee -a ${BASE_NAME}_logs.txt
-echo  "Make sure conda environment has all the necessary programs installed." | tee -a ${BASE_NAME}_logs.txt
-echo "-----------------------------------------------------------------------------------------------" | tee -a ${BASE_NAME}_logs.txt
-
 # Start pipeline
-echo "Analysis started: $(date +'%Y:%m:%d:%H:%M')" | tee -a ${BASE_NAME}_logs.txt
+BASE_NAME="${EXP_ID}_${EXP_TYPE}"
 
+show_info | tee -a ${BASE_NAME}_logs.txt
+echo "# Analysis started: $(date +'%Y:%m:%d:%H:%M')\n" | tee -a ${BASE_NAME}_logs.txt
+
+## Check if specified fastq file exists
 if [ ! -e "${BASE_NAME}.fastq.gz" ]; then
-  echo "${BASE_NAME}.fastq.gz file does not exist." | tee -a ${BASE_NAME}_logs.txt
+  echo "# ${BASE_NAME}.fastq.gz file does not exist." | tee -a ${BASE_NAME}_logs.txt
   on_exit
   exit 1
 fi
+
+## Check if demultiplexing option is on.
+if [ "$DEMUX" == "yes" ]; then
+  # Check if the BC file exists in the current directory.
+  if [ ! -e "${BASE_NAME}_BC.txt" ]; then
+  # If BC file does not exists, exit.
+    echo "# Barcode information not found! Make sure to provide proper barcode file." | tee -a ${BASE_NAME}_logs.txt
+    echo "# Barcode file should have the same name convention as fastq.gz file, followed by '_BC.txt'.\n" | tee -a ${BASE_NAME}_logs.txt
+    on_exit
+    exit 1
+  else
+    echo "# Barcode information found! Demultiplexing will be performed based on the barcode information.\n" | tee -a ${BASE_NAME}_logs.txt
+  fi
+else
+  echo "# Demultiplexing will be skipped.\n" | tee -a ${BASE_NAME}_logs.txt
+fi
+
+## Check if bedtools genome index file exists.
+if [ ! -e "${genome_index}.fa.fai" ]; then
+# If BC file does not exists, exit.
+  echo "# Genome index for bedtools not found! Make sure to provide proper genome file for bedtools.\n" | tee -a ${BASE_NAME}_logs.txt
+  on_exit
+  exit 1
+else
+  echo "# Genome index for bedtools found!\n" | tee -a ${BASE_NAME}_logs.txt
+fi
+
+## Check if bowtie2 genome index file exists.
+bt2_index=("1.bt2" "2.bt2" "3.bt2" "4.bt2" "rev.1.bt2" "rev.2.bt2")
+
+for index in "${bt2_index[@]}"; do
+  file_to_check="${genome_index}.${index}"
+  if [ ! -e "${file_to_check}" ]; then
+    echo "# ${index} index file not found! Make sure all bowtie2 genome index files are in the provided directory.\n" | tee -a ${BASE_NAME}_logs.txt
+    on_exit
+    exit 1
+  else
+    echo "# ${index} index file found!" | tee -a ${BASE_NAME}_logs.txt
+  fi
+done
 
 ## Make directories:
 echo "\n# Create necessary directories: " | tee -a ${BASE_NAME}_logs.txt
@@ -179,10 +240,15 @@ mkdir ${BASE_NAME}_peaks
 echo "\n# Unzip fasta.gz file." | tee -a ${BASE_NAME}_logs.txt
 gunzip -k ${BASE_NAME}.fastq.gz
 
+## Filter reads based on quality:
+echo "\n# Quality fiilter reads using fastx fastq_quality_filter:" | tee -a ${BASE_NAME}_logs.txt
+echo "\n  ${BASE_NAME}.fastq --> ${BASE_NAME}_filtered.fastq" | tee -a ${BASE_NAME}_logs.txt
+fastq_quality_filter -v -q ${QUALITY} -i ${BASE_NAME}.fastq -o ${BASE_NAME}_filtered.fastq | tee -a ${BASE_NAME}_logs.txt
+
 ## Collapse identical reads:
 echo "\n# Collapse identical reads using fastx_collapser:" | tee -a ${BASE_NAME}_logs.txt
-echo "\n  ${BASE_NAME}.fastq --> ${BASE_NAME}_collapsed.fa" | tee -a ${BASE_NAME}_logs.txt
-fastx_collapser -v -Q 33 -i ${BASE_NAME}.fastq -o ${BASE_NAME}_collapsed.fa | tee -a ${BASE_NAME}_logs.txt
+echo "\n  ${BASE_NAME}_filtered.fastq --> ${BASE_NAME}_collapsed.fa" | tee -a ${BASE_NAME}_logs.txt
+fastx_collapser -v -i ${BASE_NAME}_filtered.fastq -o ${BASE_NAME}_collapsed.fa | tee -a ${BASE_NAME}_logs.txt
 rm *.fastq
 
 ## Strip UMI:
@@ -191,12 +257,15 @@ echo "\n  ${BASE_NAME}_collapsed.fa --> ${BASE_NAME}_collapsed_rmBC.fa" | tee -a
 stripBarcode.pl -len 7 -v ${BASE_NAME}_collapsed.fa ${BASE_NAME}_collapsed_rmBC.fa
 rm *_collapsed.fa
 
-## Demultiplex Samples:
-echo "\n# Demultiplex samples with barcodes using fastx_barcode_splitter, allowing ${barcodeMis} mismatches:" | tee -a ${BASE_NAME}_logs.txt
-echo "\n  ${BASE_NAME}_collapsed_rmBC.fa --> demultiplexed .fa files.\n  Unmatched.fa file will be removed at the end." | tee -a ${BASE_NAME}_logs.txt
-cat ${BASE_NAME}_collapsed_rmBC.fa | fastx_barcode_splitter.pl --bcfile ${BASE_NAME}_BC.txt --bol --mismatches ${barcodeMis} --prefix "${BASE_NAME}_" --suffix ".fa" | tee -a ${BASE_NAME}_logs.txt
-rm *_collapsed_rmBC.fa
-rm *unmatched.fa
+## Perform demultiplexing if Demux option is yes:
+if [ "$DEMUX" == "yes" ]; then
+  ## Demultiplex Samples:
+  echo "\n# Demultiplex samples with barcodes using fastx_barcode_splitter, allowing ${barcodeMis} mismatches:" | tee -a ${BASE_NAME}_logs.txt
+  echo "\n  ${BASE_NAME}_collapsed_rmBC.fa --> demultiplexed .fa files.\n  Unmatched.fa file will be removed at the end." | tee -a ${BASE_NAME}_logs.txt
+  cat ${BASE_NAME}_collapsed_rmBC.fa | fastx_barcode_splitter.pl --bcfile ${BASE_NAME}_BC.txt --bol --mismatches ${barcodeMis} --prefix "${BASE_NAME}_" --suffix ".fa" | tee -a ${BASE_NAME}_logs.txt
+  rm *_collapsed_rmBC.fa
+  rm *unmatched.fa
+fi
 
 ## Barcode, Spacer and L32 trimming.
 echo "\n# Trim barcode and spacer (GGG) using fastx_trimmer, followed by L32 removal using fastx_clipper, and keeping reads with minimum length of ${minLength}:" | tee -a ${BASE_NAME}_logs.txt
@@ -229,9 +298,9 @@ mv BAM ${BASE_NAME}_mapped_reads
 
 ## Collapse PCR dups
 echo "\n# Collapse PCR duplicates from the mapped reads using CTK tag2collapse.pl." | tee -a ${BASE_NAME}_logs.txt
-echo "  From collapsed bed file, use bedtools genomecov to make bedgraph files." | tee -a ${BASE_NAME}_logs.txt
+echo "\n# From collapsed bed file, use bedtools genomecov to make bedgraph files." | tee -a ${BASE_NAME}_logs.txt
 for input_bed in *.bed; do
-    tag2collapse.pl --keep-tag-name --keep-max-score --random-linker -EM 30 --seq-error-model em-local --weight --weight-in-name  ${input_bed} $(basename "$input_bed" .bed).collapsed.bed
+    tag2collapse.pl --keep-tag-name --keep-max-score --random-barcode -EM 30 --seq-error-model em-local -weight --weight-in-name  ${input_bed} $(basename "$input_bed" .bed).collapsed.bed
     mv ${input_bed} BED
     bedtools genomecov -i $(basename "$input_bed" .bed).collapsed.bed -strand + -g ${genome_index}.fa.fai -bg > $(basename "$input_bed" .bed).collapsed.pos.bedgraph
     bedtools genomecov -i $(basename "$input_bed" .bed).collapsed.bed -strand - -g ${genome_index}.fa.fai -bg > $(basename "$input_bed" .bed).collapsed.rev.bedgraph
